pyproject.toml
[tool.poetry]
name = "brain-agriculture"
version = "0.1.0"
description = ""
authors = ["Gustavo Santiago <gustavoalcantarasantiago@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = {extras = ["all"], version = "^0.116.1"}
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
pydantic = ">=2.11.7,<3.0.0"
python-dotenv = "^1.1.1"
validate-docbr = "^1.11.1"
pytest = "^8.4.1"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "uvicorn --workers 4 --host 0.0.0.0 --port 7001 main:app"
